requirements.txt
numpy
Pillow
fire==0.3.1
nibabel
temp
python-time
matplotlib
monai
pyyaml
matplotlib
